convert util datatype

+++ src/include/solvers/solver1.h
@@ -7,8 +7,8 @@ class Solver1 : public Solver {
     int to;
   };
 
-  std::vector<int> die_max_util;
-  std::vector<int> die_util;
+  std::vector<double> die_max_util;
+  std::vector<double> die_util;
   std::vector<std::vector<line_segment>> horizontal_contours;
   std::vector<std::vector<std::vector<std::pair<int, int>>>> spared_rows;
   int die_size;
